+++ native/c/zowex.cpp
@@ -570,9 +570,14 @@ int main(int argc, char *argv[])
   ZCLIOption uss_file_mode("mode");
   uss_file_mode.set_required(false);
   uss_file_mode.set_description("permissions");
-  ZCLIOption uss_file_attributes("attributes");
-  uss_file_attributes.set_required(false);
-  uss_file_attributes.set_description("whether to list attributes");
+  ZCLIOption uss_file_list_all("all");
+  uss_file_list_all.get_aliases().push_back("-a");
+  uss_file_list_all.set_required(false);
+  uss_file_list_all.set_description("whether to list hidden files (default: false)");
+  ZCLIOption uss_file_list_long_format("long");
+  uss_file_list_long_format.get_aliases().push_back("-l");
+  uss_file_list_long_format.set_required(false);
+  uss_file_list_long_format.set_description("whether to display the long format in list output (including permissions, ownership, etc.) (default: false)");
   ZCLIOption uss_recursive("recursive");
   uss_recursive.get_aliases().push_back("-r");
   uss_recursive.set_required(false);
@@ -596,7 +601,8 @@ int main(int argc, char *argv[])
   uss_list.set_description("list USS files and directories");
   uss_list.set_zcli_verb_handler(handle_uss_list);
   uss_list.get_positionals().push_back(uss_file_path);
-  uss_list.get_options().push_back(uss_file_attributes);
+  uss_list.get_options().push_back(uss_file_list_all);
+  uss_list.get_options().push_back(uss_file_list_long_format);
   uss_group.get_verbs().push_back(uss_list);
 
   ZCLIVerb uss_view("view");
@@ -1862,7 +1868,10 @@ int handle_uss_list(ZCLIResult result)
 
   ZUSF zusf = {0};
   string response;
-  rc = zusf_list_uss_file_path(&zusf, uss_file, response, result.get_option_value("--attributes") == "true");
+  ListOptions options;
+  options.all = result.get_option_value("--all") == "true";
+  options.long_format = result.get_option_value("--long") == "true";
+  rc = zusf_list_uss_file_path(&zusf, uss_file, response, options);
   if (0 != rc)
   {
     cerr << "Error: could not list USS files: '" << uss_file << "' rc: '" << rc << "'" << endl;