+++ native/CHANGELOG.md
@@ -7,6 +7,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## Recent Changes
 
 - `golang`: `zowed` now returns the UNIX permissions for each item's `mode` property in a USS list response. [#341](https://github.com/zowe/zowe-native-proto/pull/341)
+- `c`: Added `--long` option to the `zowex uss list` command to return the long format for list output, containing ownership, permissions and additional file metadata. [#341](https://github.com/zowe/zowe-native-proto/pull/341)
+- `c`: Added `--all` option to the `zowex uss list` command to show hidden files when listing a UNIX directory. [#341](https://github.com/zowe/zowe-native-proto/pull/341)
 
 ## `0.1.2`